style - moved wrapped data loader to its own file, data rader to data loaders
