+++ pkg/storage/loaders/operational_needs_loader.go
@@ -14,7 +14,7 @@ import (
 // GetOperationalNeedsByModelPlanID uses a data loader to aggregate SQL calls and return data
 func (loaders *DataLoaders) GetOperationalNeedsByModelPlanID(ctx context.Context, keys dataloader.Keys) []*dataloader.Result {
 	logger := appcontext.ZLogger(ctx)
-	arrayCK, err := CompoundKeyArray(keys)
+	arrayCK, err := ConvertToKeyArgsArray(keys)
 	if err != nil {
 		logger.Error("issue converting keys for data loader in Operational Solutions", zap.Error(*err))
 	}
